Css Updates
Changed the ph_plays.html to index.html. Added links to the ph_styles_txt.css and the Ph_layout.css. so now the styles and the layout is going to perform on the index.html file. Added the font face rule with the fonts that were required. Put woff format for the newer web pages and the ttf for the older formats.The style rule is for the html and it is saying that the background color is 91 hue, has 8% saturation, 56% of lightness. The style for the header is saying that the Background-Color for the header is going to be Black. For all of the paragraphs, they will have a margin of 0px, and they will have a padding on 5 pixels on top and 25 pixels on the right,bottom, and left. The style rule is saying that all of the paragraphs that are direct child of the body will have a font-size of 1.1em and a horizontal center.The address style rule is that the font-style will be normal, the size will be 0.9em, a horizontal center and a padding of 10 pixels on top and thats it. The color is the color that they text is going to be. The font-family is where every line line of text in the nav is going to be the  ‘Trebuchet MS’, Helvetica, sans-serif font. The padding is just stating that there will be 10 pixels on the top and no more spacing adding on the right, bottom, and left.This style rule is for any links that are visited, the links are white, the links do not have underlines or any type of decoration and the background color of the links are semi-transparent. This style rule is for any links that are clicked or someone hovers the link, the color is a grey and the background color is a semi-transparent.