Final CSS commit
This id is for the first section and the rule is saying that the background-color is going to be hsl(240, 100%, 88%). This id is for the second section and the rule is saying that the background-color is going to be hsl(25, 88%, 73%). This id is for the third section and the rule is saying that the background-color is going to be hsl(0, 100%, 75%). This id id for the four section and the rule is saying that the background-color is going to be hsl(296, 86%, 86%).  This section id is for the primary headings in the first play section and the style rule is saying that the font family is champagne, cursive. This section id is for the primary headings in the second play section and the style rule is saying that the font family is Grunge, ‘Times New Roman’, Times, serif. This section id is for the primary headings in the third play section and the style rule is saying that the font family is Impact, Charcoal, sans-serif. This section id is for the primary headings in the fourth play section and the style rule is saying that the font family is Dobkin, cursive. The style rule is for the dt element and the rule is saying that the font size is 1.3em, the wieght is bold and the color hsla(0, 0%, 0%, 0.4). The style rule is for the dt element and the rule is saying that the font size is 1.3em, margin left 0 pixels and the margin-bottom is 10 pixels.